Refactored afterSuccess into formatOutput.js and exposed it as part of utils
