+++ lib/bindUtils/index.js
@@ -2,11 +2,13 @@
 const processBody = require('../rawHttpRequest/processBody');
 const validateUrlInput = require('../rawHttpRequest/validateUrlInput');
 const validateRequestAgainstProtectedService = require('../rawHttpRequest/validateRequestAgainstProtectedService.js');
+const formatOutput = require('../rawHttpRequest/formatOutput');
 function setupRawHttpRequestUtils () {
 	falafel.utils.rawHttpRequest = {
-		processBody,
 		validateUrlInput,
-		validateRequestAgainstProtectedService
+		processBody,
+		validateRequestAgainstProtectedService,
+		formatOutput
 	};
 }
