+++ lib/rawHttpRequest/rawHttpRequestModel.js
@@ -3,6 +3,7 @@ const _ = require('lodash');
 const processBody = require('./processBody.js');
 const validateUrlInput = require('./validateUrlInput.js');
 const validateRequestAgainstProtectedService = require('./validateRequestAgainstProtectedService.js');
+const formatOutput = require('./formatOutput.js');
 
 function convertArrayFormatToObject (targetArray) {
 	return _.reduce(targetArray, (acc, { key, value }) => {
@@ -12,11 +13,11 @@ function convertArrayFormatToObject (targetArray) {
 
 module.exports = {
 
-	globals: true,
+	globals: true, //Use global model
 
 	before: (params) => {
 
-		validateUrlInput(params);
+		validateUrlInput(params); //Validate `endpoint` and `full_url`
 
 		return processBody(params.body)
 		.then((body) => {
@@ -61,26 +62,13 @@ module.exports = {
 
 	data: '{{processedBody}}',
 
+	/*
+		Performs validation of the request against protected service whitelist
+		if enabled and configured
+	*/
 	beforeRequest: validateRequestAgainstProtectedService,
 
-	afterSuccess: (body, params, res) => {
-		if (body instanceof Buffer) {
-			body = body.toString('utf8');
-			try {
-				body = JSON.parse(body);
-			} catch (e) {
-				//Continue as normal
-			}
-		}
-		const result = {
-			status_code: res.status_code,
-			headers: res.headers,
-			body: body
-		};
-		if (params.include_raw_body) {
-			result.raw_body = res.raw.toString();
-		}
-		return result;
-	}
+	//Transform the API response into a standardised format
+	afterSuccess: formatOutput
 
 };
